requirements.txt
fastapi
uvicorn
jinja2
python-multipart
playwright
google-genai
python-dotenv
# tomli is only needed for python < 3.11
tomli; python_version < "3.11"
httpx
beautifulsoup4
